For soem reason my last commit was all kinds of messed up, so I cleaned up the commit and re-rolled it through this commit.
Changelog: This is a re-roll of the _checkPropertyStatus so that we have less server intensive checks and also that it is more intelligent to the feed coming through.

Also, this commit addresses some code formatting issues and also the messages reported in the admin section to make them less confusing to the end-user.